+++ src/main/kotlin/de/keeyzar/gpthelper/gpthelper/features/setup/domain/service/YamlModificationService.kt
@@ -3,4 +3,5 @@ package de.keeyzar.gpthelper.gpthelper.features.setup.domain.service
 interface YamlModificationService {
     fun addDependency(yamlContent: String, dependencyName: String, dependencyValue: Any): String
     fun addFlutterGenerate(yamlContent: String): String
+    fun createL10nYaml(l10nConfigMap: Map<String, Any>): String
 }